Move Params struct to new files
